refactor (schema-export): Simplify schema-export file structure
